feat: refactor constants for event list
