+++ constants/eventList.ts
@@ -1,13 +1,4 @@
-export interface eventsList {
-  month: string;
-  events: eventsItem[];
-}
-
-export interface eventsItem {
-  title: string;
-  description: string;
-  date: string;
-}
+import { eventsList } from "@/types/eventTypes";
 
 export const EventsList: eventsList[] = [
   {
