+++ types/eventTypes.d.ts
@@ -1,0 +1,10 @@
+export interface eventsList {
+  month: string;
+  events: eventsItem[];
+}
+
+export interface eventsItem {
+  title: string;
+  description: string;
+  date: string;
+}
